requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
python-dotenv==1.0.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5 
google-generativeai